Refact build segment logic
